requirements.txt
pandas>=2.0.0
Jinja2>=3.1.0
rich>=13.0.0
python-dateutil>=2.8.2
